pyproject.toml
[tool.poetry]
name = "jb-engine"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^1.3.4"
tenacity = "^8.2.3"
termcolor = "^2.3.0"
transitions = "^0.9.0"
lark = "^1.1.8"
aiohttp = "^3.9.0"
black = "^24.1.1"
asyncio = "^3.4.3"
requests = "^2.31.0"
azure-storage-blob = "^12.19.0"
confluent-kafka = "^2.3.0"
bs4 = "^0.0.2"
pyaml = "^23.12.0"
jsonpickle = "^3.0.2"
opencensus-ext-azure = "^1.1.13"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

pre-commit = "^3.7.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.black]
line-length = 88
